requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
ultralytics==8.0.92
scikit-learn==1.2.2
torch==2.1.0
torchvision==0.16.0
Pillow==10.0.1
numpy==1.24.3
joblib==1.3.2
tqdm==4.66.1
